+++ packages/in_app_purchase/in_app_purchase_platform_interface/lib/src/types/types.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'in_app_purchase_error.dart';
 export 'product_details.dart';
 export 'product_details_response.dart';
 export 'purchase_details.dart';