+++ packages/video_player/video_player_platform_interface/test/method_channel_video_player_test.dart
@@ -92,10 +92,12 @@ class _ApiLogger implements TestHostVideoPlayerApi {
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
+  // Store the initial instance before any tests change it.
+  final VideoPlayerPlatform initialInstance = VideoPlayerPlatform.instance;
+
   group('$VideoPlayerPlatform', () {
     test('$MethodChannelVideoPlayer() is the default instance', () {
-      expect(VideoPlayerPlatform.instance,
-          isInstanceOf<MethodChannelVideoPlayer>());
+      expect(initialInstance, isInstanceOf<MethodChannelVideoPlayer>());
     });
   });
 
@@ -232,13 +234,16 @@ void main() {
     });
 
     test('videoEventsFor', () async {
-      ServicesBinding.instance?.defaultBinaryMessenger.setMockMessageHandler(
+      _ambiguate(ServicesBinding.instance)
+          ?.defaultBinaryMessenger
+          .setMockMessageHandler(
         "flutter.io/videoPlayer/videoEvents123",
         (ByteData? message) async {
           final MethodCall methodCall =
               const StandardMethodCodec().decodeMethodCall(message);
           if (methodCall.method == 'listen') {
-            await ServicesBinding.instance?.defaultBinaryMessenger
+            await _ambiguate(ServicesBinding.instance)
+                ?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -250,7 +255,8 @@ void main() {
                     }),
                     (ByteData? data) {});
 
-            await ServicesBinding.instance?.defaultBinaryMessenger
+            await _ambiguate(ServicesBinding.instance)
+                ?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -259,7 +265,8 @@ void main() {
                     }),
                     (ByteData? data) {});
 
-            await ServicesBinding.instance?.defaultBinaryMessenger
+            await _ambiguate(ServicesBinding.instance)
+                ?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -272,7 +279,8 @@ void main() {
                     }),
                     (ByteData? data) {});
 
-            await ServicesBinding.instance?.defaultBinaryMessenger
+            await _ambiguate(ServicesBinding.instance)
+                ?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -281,7 +289,8 @@ void main() {
                     }),
                     (ByteData? data) {});
 
-            await ServicesBinding.instance?.defaultBinaryMessenger
+            await _ambiguate(ServicesBinding.instance)
+                ?.defaultBinaryMessenger
                 .handlePlatformMessage(
                     "flutter.io/videoPlayer/videoEvents123",
                     const StandardMethodCodec()
@@ -325,3 +334,10 @@ void main() {
     });
   });
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+// TODO(ianh): Remove this once we roll stable in late 2021.
+T? _ambiguate<T>(T? value) => value;