+++ packages/google_sign_in/google_sign_in/CHANGELOG.md
@@ -1,6 +1,26 @@
-## NEXT
+## 5.1.1
+
+* Update minimum Flutter SDK to 2.5 and iOS deployment target to 9.0.
+
+## 5.1.0
+
+* Add reAuthenticate option to signInSilently to allow re-authentication to be requested
+
+* Updated Android lint settings.
+
+## 5.0.7
+
+* Mark iOS arm64 simulators as unsupported.
+
+## 5.0.6
+
+* Remove references to the Android V1 embedding.
+
+## 5.0.5
 
 * Add iOS unit and UI integration test targets.
+* Add iOS unit test module map.
+* Exclude arm64 simulators in example app.
 
 ## 5.0.4